style code check

+++ modules/comb_gen/src/Comb_gen.cpp
@@ -25,12 +25,12 @@ void CombGen::CombGenerate() {
   std::vector<int> combErr;
   auto iter = combErr.cbegin();
   combErr.insert(iter, this->K, -1);
-  for (int i = 0; i < comb.size(); i++) {
+  for (long unsigned int i = 0; i < comb.size(); i++) {
     comb[i] = 0;
   }
   while (comb != combErr) {
     std::vector<int> combCopy;
-    for (int i = 0; i < comb.size(); i++) {
+    for (long unsigned int i = 0; i < comb.size(); i++) {
       combCopy.push_back(VecIn[comb[i]]);
     }
     VecOut.push_back(combCopy);
